Added conversion from Fahrenheit to Celsius and vice versa
Removed temp max and min in HTML. Added IDs to the links in HTML to be selected from JS. Created global variables in JS for event listeners. Created celsiusTemperature variable which gets overwritten every time API is called. Created event listener functions that change UI when clicked triggering the active class in CSS to correspond to the current active link.